// apps/erp/src/features/projects/types.ts
export type ClientType = "b2b" | "b2c";

export interface ProjectClient {
  id: string;
  name_ar: string;
  name_en: string;
  type: ClientType;
  trn: string | null;
}